requirements.txt
functions-framework==3.*
requests==2.28.2
python-decouple==3.7